feat(streaming): add WebSocket real-time viewer updates handler
## Overview
Implemented critical WebSocket support for real-time streaming metrics.
This unblocks P1 viewer story by enabling live viewer count updates.

## Changes

### 1. New WebSocket Handler Module
- Created `backend/user-service/src/handlers/streaming_websocket.rs` (~260 lines)
  - StreamingHub actor for connection management
  - StreamingWebSocket actor for per-connection handling
  - WebSocket upgrade handler: GET /api/v1/streams/{stream_id}/ws
  - Helper functions for broadcasting updates

### 2. Broadcast Message Protocol
All messages use JSON format with event types:
- `viewer_count_changed` - When viewer count updates
- `stream_started` - When broadcast begins
- `stream_ended` - When broadcast concludes
- `quality_changed` - When bitrate adaptation occurs

### 3. Architecture
- Hub-spoke pattern: Central StreamingHub broadcasts to all connected clients
- Session tracking: Each connection gets unique ID in hub
- Stateless: Hub uses in-memory map, survives restarts via reconnect
- Non-blocking: Uses actix async actors, no database blocking

### 4. Integration Points
- Added StreamingHub to app state in main.rs
- Registered WebSocket route: GET /api/v1/streams/{stream_id}/ws
- Helper functions exported for services to trigger broadcasts:
  - notify_viewer_count_changed()
  - notify_stream_started()
  - notify_stream_ended()

### 5. Dependencies Added
- actix 0.13 (actor framework)
- actix-web-actors 4.3 (WebSocket support)

## Alignment with Spec
Resolves critical gap identified in CODE_ALIGNMENT.md:
- ✅ WebSocket real-time hub (T050-T052)
- ✅ Enables live viewer count updates (P1 requirement)
- ✅ Enables stream status notifications

## Testing Notes
- Unit tests for message serialization included
- Integration tests will be added in next phase
- Requires Redis connection manager in app state (already present)

## Next Steps (Phase 2)
1. Add integration tests with mock RTMP client
2. Implement Prometheus metrics exporter
3. Complete API documentation
4. Write deployment guide

## Technical Debt
- Connection cleanup on server restart (graceful shutdown)
- Per-stream rate limiting on broadcasts
- Memory monitoring for large viewer counts

+++ backend/user-service/src/redis/operations.rs
@@ -24,7 +24,7 @@ pub async fn redis_set_ex(
 ) -> Result<()> {
     let mut conn = redis.clone();
     let _: () = conn
-        .set_ex(key, value, ttl_seconds as usize)
+        .set_ex(key, value, ttl_seconds)
         .await
         .map_err(|e| anyhow!("Failed to set Redis key '{}': {}", key, e))?;
     Ok(())
@@ -147,7 +147,7 @@ pub async fn redis_expire(
 ) -> Result<bool> {
     let mut conn = redis.clone();
     let expired: bool = conn
-        .expire(key, ttl_seconds as usize)
+        .expire(key, ttl_seconds as i64)
         .await
         .map_err(|e| anyhow!("Failed to set expiration on Redis key '{}': {}", key, e))?;
     Ok(expired)
